pyproject.toml
[project]
name = "receptai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faster-whisper>=1.2.1",
    "google-genai>=1.68.0",
    "pandas>=2.2.0",
    "python-dotenv>=1.2.2",
    "scikit-learn>=1.8.0",
    "sounddevice>=0.5.5",
    "streamlit>=1.45.0",
    "audio-recorder-streamlit>=0.0.10",
    "python-docx>=1.2.0",
]
